feat(tj-learning): 完成我的课表相关接口功能 - 创建 LessonChangeListener 消息监听器处理订单支付和退款 - 订单支付后监听MQ消息并将课程加入我的课表 - 实现查询全部我的全部课程、正在学习课程等功能 - 实现按课程id删除课程，及退款后监听MQ消息并删除相关课程

+++ tj-trade/src/main/java/com/tianji/trade/service/impl/CartServiceImpl.java
@@ -83,7 +83,7 @@ public class CartServiceImpl extends ServiceImpl<CartMapper, Cart> implements IC
     }
 
     private void checkCartsFull(Long userId) {
-        int count = lambdaQuery().eq(Cart::getUserId, userId).count();
+        long count = lambdaQuery().eq(Cart::getUserId, userId).count();
         if (count >= tradeProperties.getMaxCourseAmount()) {
             throw new BizIllegalException(
                     StringUtils.format(CARTS_FULL, tradeProperties.getMaxCourseAmount()));
@@ -91,7 +91,7 @@ public class CartServiceImpl extends ServiceImpl<CartMapper, Cart> implements IC
     }
 
     private boolean checkCourseExists(Long courseId, Long userId) {
-        int count = lambdaQuery()
+        long count = lambdaQuery()
                 .eq(Cart::getUserId, userId)
                 .eq(Cart::getCourseId, courseId)
                 .count();
